backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
motor==3.6.0
redis==5.0.0
aiokafka==0.10.0
neo4j==5.23.0
PyJWT==2.9.0
bcrypt==4.2.0
pydantic==2.9.0
python-multipart==0.0.9
